requirements.txt
accelerate==1.7.0
safetensors==0.5.3
xformers==0.0.30
runpod==1.7.9
torch==2.7.0
torchvision
Pillow
numpy>=1.24.0
scipy
triton
huggingface-hub
hf_transfer
hf_xet
setuptools
pruna==0.2.5
